pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-picrawler"
version = "0.1.0"
description = "MCP server that gives Claude a body on a SunFounder PiCrawler"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Guy" }]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.27",
    "pillow>=10.0",
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
]

[project.optional-dependencies]
pi = [
    "picrawler",
    "robot-hat",
    "vilib",
]
vision = [
    "ultralytics>=8.2",
]
caption = [
    "transformers>=4.40",
    "torch>=2.2",
    "accelerate>=0.30",
    "einops>=0.7",
    "timm>=0.9",
]
voice = [
    "openai>=1.50",
    "sounddevice>=0.4.6",
    "websockets>=12",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
mcp-picrawler = "mcp_picrawler.__main__:main"
picrawler-voice = "voice_agent.agent:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_picrawler", "src/voice_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    # mcp SDK is nudging us toward streamable_http_client (different signature).
    # Suppress until we migrate — see voice_agent/mcp_bridge.py comment.
    "ignore:Use `streamable_http_client` instead.:DeprecationWarning",
]
